Redirect while api client is not ready #32
In order to interact with the views, we need a client in ready state.

Even thought views required a client are hidden from the UI, a user could navigate to them via URL. In that case the page silently fails.

Instead, we should detect whether the client is ready (See `~/src/vue/pages/Index.vue` where `ready` is used) in a router hook and redirect them to `/services` when they are not already in a `/services` subroute.