feat: Add eslint linting stage
The pipeline's build step will fail if the code is not linted correctly, however this is not really descriptive. Instead, we should add linting to the unit test job, failing it. This will produce a more meaningful error, earlier in the pipeline.

Add eslint to the unit test job step.

Closes #133